+++ src/pgpack/common/__init__.py
@@ -10,8 +10,8 @@ from . import (
 from .casts import pandas_astype
 from .detector import detect_oid
 from .metadata import (
+    PGPackMeta,
     metadata_from_frame,
-    metadata_reader,
 )
 from .param import PGParam
 from .repr import (
@@ -24,13 +24,13 @@ from .repr import (
 __all__ = (
     "Error",
     "Fmt",
+    "PGPackMeta",
     "PGParam",
     "Signature",
     "Size",
     "compile_pgtype",
     "detect_oid",
     "metadata_from_frame",
-    "metadata_reader",
     "pandas_astype",
     "pgpack_repr",
     "table_repr",